requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
cs50==9.4.0
Flask==3.1.0
Flask-Session==0.8.0
greenlet==3.1.1
hiredis==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
msgspec==0.19.0
packaging==24.2
python-dotenv==1.0.1
redis==5.2.1
SQLAlchemy==2.0.37
sqlparse==0.5.3
termcolor==2.5.0
typing_extensions==4.12.2
Werkzeug==3.1.3
wheel==0.45.1
